@@ -1,6 +1,6 @@
 {
   "name": "oerc-client",
-  "version": "0.10.2",
+  "version": "0.11.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -12,15 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.0.3",
-    "@angular/common": "^14.0.3",
-    "@angular/compiler": "^14.0.3",
-    "@angular/core": "^14.0.3",
-    "@angular/forms": "^14.0.3",
-    "@angular/localize": "^14.0.3",
-    "@angular/platform-browser": "^14.0.3",
-    "@angular/platform-browser-dynamic": "^14.0.3",
-    "@angular/router": "^14.0.3",
+    "@angular/animations": "^15.0.4",
+    "@angular/common": "^15.0.4",
+    "@angular/compiler": "^15.0.4",
+    "@angular/core": "^15.0.4",
+    "@angular/forms": "^15.0.4",
+    "@angular/localize": "^15.0.4",
+    "@angular/platform-browser": "^15.0.4",
+    "@angular/platform-browser-dynamic": "^15.0.4",
+    "@angular/router": "^15.0.4",
     "@egjs/hammerjs": "^2.0.17",
     "flatpickr": "^4.6.13",
     "keycharm": "^0.4.0",
@@ -38,20 +38,20 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.0.3",
-    "@angular-eslint/builder": "14.0.3",
-    "@angular-eslint/eslint-plugin": "14.0.3",
-    "@angular-eslint/eslint-plugin-template": "14.0.3",
-    "@angular-eslint/schematics": "14.0.3",
-    "@angular-eslint/template-parser": "14.0.3",
-    "@angular/cli": "^14.0.3",
-    "@angular/compiler-cli": "^14.0.3",
-    "@types/jasmine": "^3.10.6",
+    "@angular-devkit/build-angular": "^15.0.4",
+    "@angular-eslint/builder": "15.1.0",
+    "@angular-eslint/eslint-plugin": "15.1.0",
+    "@angular-eslint/eslint-plugin-template": "15.1.0",
+    "@angular-eslint/schematics": "15.1.0",
+    "@angular-eslint/template-parser": "15.1.0",
+    "@angular/cli": "^15.0.4",
+    "@angular/compiler-cli": "^15.0.4",
+    "@types/jasmine": "^4.3.1",
     "@types/node": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "5.29.0",
-    "@typescript-eslint/parser": "5.29.0",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
     "codelyzer": "^6.0.2",
-    "eslint": "^8.15.0",
+    "eslint": "^8.28.0",
     "jasmine-core": "^4.1.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "^6.3.19",
@@ -62,7 +62,7 @@
     "postcss": "^8.4.13",
     "protractor": "~7.0.0",
     "ts-node": "~10.9.1",
-    "typescript": "~4.7.4"
+    "typescript": "~4.8.4"
   },
   "optionalDependencies": {
     "fsevents": "^2.3.2"
