@@ -1,4 +1,3 @@
-//
 // oerc, alias oer-collector
 // Copyright (C) 2021 emschu[aet]mailbox.org
 //
@@ -37,6 +36,8 @@ func StartServer() {
 		gin.SetMode("release")
 	}
 
+	go setupMaterializedView()
+
 	r := initRouter()
 	log.Printf("Starting API server...\n")
 	ip := net.ParseIP(GetAppConf().ServerHost)
@@ -51,6 +52,44 @@ func StartServer() {
 		log.Printf("Problem starting server %v\n", err)
 		return
 	}
+}
+
+func setupMaterializedView() {
+	db := getDb()
+	db.Exec(`drop materialized view status_info`)
+	db.Exec(`create materialized view IF NOT EXISTS status_info as
+SELECT min(program_entries.start_date_time) AS data_start_time,
+       max(program_entries.end_date_time)   AS data_end_time,
+       count(*) as program_entry_count,
+       (SELECT count(*) from channel_families) as channel_family_count,
+       (SELECT count(*) from channels) as channel_count,
+       (SELECT count(*) from image_links) as image_link_count,
+       (SELECT count(*) from tv_shows) as tv_show_count,
+       (SELECT count(*) from log_entries) as log_count,
+       (SELECT count(*) from recommendations) as recommendation_count,
+       now() as created_at
+FROM program_entries`)
+	if isDebug() {
+		log.Printf("Materialized status view")
+	}
+}
+
+// StatusInfoModel a gorm model for the materialized view
+type StatusInfoModel struct {
+	ChannelFamilyCount  uint64
+	ChannelCount        uint64
+	ProgramEntryCount   uint64
+	TvShowCount         uint64
+	ImageLinkCount      uint64
+	LogCount            uint64
+	RecommendationCount uint64
+	DataStartTime       string
+	DataEndTime         string
+}
+
+// TableName of the materialized view prefixed by db schema
+func (s *StatusInfoModel) TableName() string {
+	return fmt.Sprintf("%s.status_info", appConf.DbSchema)
 }
 
 // StatusResponse api object
